pyproject.toml
[tool.poetry]
name = "decoupled-dj"
version = "0.1.0"
description = "Follows Valentino Gagliardi's Decoupled Django"
authors = ["Preston Mackert"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.4"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.5"
uvicorn = "^0.20.0"
djangorestframework = "^3.14.0"
django-simple-history = "^3.2.0"
django-auditlog = "^2.2.1"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]
django-extensions = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
